requirements.txt
requests
pillow
redis